requirements.txt
black==22.3.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.0
flake8==4.0.1
idna==3.3
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
pycodestyle==2.8.0
pyflakes==2.4.0
pyTelegramBotAPI==4.4.0
requests==2.27.1
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.9
